fix: changed the storage on the backend and added a new viewed query

+++ frontend/src/app/shared/interceptors/auth.interceptor.ts
@@ -1,26 +1,37 @@
-import type {HttpInterceptorFn} from '@angular/common/http';
+import type {HttpErrorResponse, HttpInterceptorFn} from '@angular/common/http';
 import {LocalStorageService} from '../services/local-storage.service';
 import {inject} from '@angular/core';
 import {tap} from 'rxjs';
 import {Router} from '@angular/router';
 import {AppRoutesEnum} from '../enums/app-router.enum';
+import {environment} from '../../../environments/environment';
+
 
 export const authInterceptorFunction: HttpInterceptorFn = (request, next) => {
   const localStorageService = inject(LocalStorageService);
   const accessToken = localStorageService.getRefreshToken();
   const router = inject(Router);
 
-  if(accessToken) {
+  const myRequest = request.url.startsWith(environment.api);
+  console.log(myRequest)
+  console.log(request.url)
+
+  if (accessToken && myRequest) {
+    console.log(accessToken)
     const newRequest = request.clone({
       setHeaders: {'x-auth': accessToken},
     })
 
     return next(newRequest).pipe(
       tap({
-        error: () => {
-          console.log('ошибка в интерcепторe')
-          localStorageService.removeTokens();
-          router.navigate(['/' + AppRoutesEnum.MAIN])
+        error: (error: HttpErrorResponse) => {
+          console.log('ошибка в интерcепторe');
+          console.log(error);
+          if (error.status === 401) {
+            console.log('ошибка в интерcепторe 401')
+            localStorageService.removeTokens();
+            router.navigate(['/' + AppRoutesEnum.MAIN])
+          }
         }
       })
     );
